Validated by the complete GitHub Quality workflow, including clean dependency installation, frontend checks, 276 regular backend tests, and real-MySQL migration and concurrency tests.

+++ backend/Repositories/InventoryEventRepository.php
@@ -39,7 +39,7 @@ class InventoryEventRepository implements RepositoryInterface
         throw new \BadMethodCallException('Method not supported on log repository.');
     }
 
-    public function record(int $productId, string $action, int $newQuantity, int $delta = 0): void
+    public function record(int $productId, string $action, float $newQuantity, float $delta = 0.0): void
     {
         $this->model->record($productId, $action, $newQuantity, $delta);
     }
